modularization of routes and fix index.js
