+++ controllers/category/categoriesFunctions.js
@@ -1,7 +1,7 @@
-const categoryRouter = require('express').Router();
-const Category = require('../models/products/Category')
+const Category = require('../../models/products/Category')
 
-categoryRouter.get('/', async (req, res)=>{
+
+const getCategories = async (req, res)=>{
     await Category.find({})
         .then(category => {
             const response = category.map(cate => {
@@ -13,9 +13,9 @@ categoryRouter.get('/', async (req, res)=>{
             res.send(err);
         })
 
-})
+}
 
-categoryRouter.get('/:name', async (req, res) => {
+const getCategoriesByName = async (req, res) => {
     const { name } = req.params;
     const match = await Category.findOne({name});
 
@@ -31,9 +31,9 @@ categoryRouter.get('/:name', async (req, res) => {
     
         res.status(200).send(responseMap);
     }
-})
+}
 
-categoryRouter.post('/create', async (req, res) => {
+const createCategory = async (req, res) => {
     const categoryToCreate = req.body;
 
     if(!categoryToCreate){
@@ -49,44 +49,19 @@ categoryRouter.post('/create', async (req, res) => {
         console.log(err)
     })
     
-})
+}
 
-categoryRouter.delete('/delete/:name', async (req, res) => {
+const deleteCategory = async (req, res) => {
     const { name } = req.params;
 
     await Category.deleteOne({name})
         .then(category => res.send(category))
         .catch(err => console.log(`Error deleting Product: `, err))
-})
+}
 
-    
-
-module.exports = categoryRouter;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// Category.findOne({ name })
-        // .then(products =>{
-        //     if(products) {
-        //         return res.json(products)
-        //     } else {
-        //         res.status(404).end()
-        //     }
-        // })// busca nota por id mas facil xd
+module.exports = {
+    getCategories,
+    getCategoriesByName,
+    createCategory,
+    deleteCategory
+}
